@@ -9,9 +9,9 @@
     "config": {
         "sort-packages": true
     },
-    "version": "100.3.4",
+    "version": "100.3.5",
     "require": {
-        "php": "~7.1.3||~7.2.0||~7.3.0",
+        "php": "~7.3.0||~7.4.0",
         "magento/framework": "102.0.*",
         "magento/module-catalog": "103.0.*",
         "magento/module-catalog-inventory": "100.3.*",
